[utils] Rename logging to logmanager
