+++ align/__init__.py
@@ -1,7 +1,7 @@
 __version__ = '0.9.8'
 
-from .utils import logging
-logging.configure_logging()
+from .utils import logmanager
+logmanager.configure_logging()
 
 from .main import schematic2layout
 from .cmdline import CmdlineParser
